error画面の追加

+++ app/Http/Controllers/DiaryController.php
@@ -60,6 +60,11 @@ class DiaryController extends Controller
         //Diaryモデルを使用してIDが一致する日記の取得
         $diary = Diary::find($id);
 
+        //ログインユーザーが日記の投稿者かチェックする
+        if (Auth::user()->id != $diary->user_id){
+            abort(403);
+        }
+        
         //取得した日記の削除
         $diary->delete();
 
@@ -67,10 +72,14 @@ class DiaryController extends Controller
         return redirect()->route('diary.index');
     }
     //編集画面を表示する
-    public function edit(int $id)
+    public function edit(Diary $diary)
     {
+        //ログインユーザーが日記の投稿者かチェックする
+        if (Auth::user()->id != $diary->user_id){
+            abort(403);
+        }
         //受け取ったIDをもとに日記を取得
-        $diary = Diary::find($id);
+        // $diary = Diary::find($id);
 
         // 編集画面を返す。同時に画面に取得した日記を渡す
         return view('diaries.edit',[
@@ -87,6 +96,11 @@ class DiaryController extends Controller
         // dd($request->title);
         //受け取ったIDをもとに日記を取得
         $diary = Diary::find($id);
+
+         //ログインユーザーが日記の投稿者かチェックする
+         if (Auth::user()->id != $diary->user_id){
+            abort(403);
+        }
 
         //取得した日記のタイトル、本文を置き換える
         $diary->title = $request->title;
